+++ src/main/java/nextstep/domain/TimeTable.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.domain;
 
 import java.time.LocalTime;
